+++ src/main/java/com/brand/blockus/Blockus.java
@@ -7,6 +7,7 @@ import com.brand.blockus.itemgroups.BlockusItemGroups;
 import com.brand.blockus.itemgroups.content.*;
 import com.brand.blockus.utils.effect.BlockusEffects;
 import com.brand.blockus.worldgen.BlockusWorldgenFeatures;
+import com.brand.blockus.worldgen.foliage.BlockusFoliagePlacerType;
 import com.google.common.reflect.Reflection;
 import net.fabricmc.api.ModInitializer;
 import net.minecraft.text.Text;
@@ -32,6 +33,7 @@ public class Blockus implements ModInitializer {
         Reflection.initialize(BlockusItems.class);
         Reflection.initialize(BlockusEntities.class);
         Reflection.initialize(BlockusEffects.class);
+        Reflection.initialize(BlockusFoliagePlacerType.class);
 
         BlockusItemGroups.init();
         BuildingBlocksGroup.init();